repositories: move generic methods to a mongodb base repo class
