+++ byoeb-v1/byoeb/byoeb/repositories/mongodb_message_repository.py
@@ -1,77 +1,10 @@
 """MongoDB implementation of MessageRepository."""
-from typing import List, Dict, Any, Optional, Tuple
-from pymongo import UpdateOne
-from pymongo.asynchronous.collection import AsyncCollection
+from typing import List, Dict, Any, Optional
 from byoeb.repositories.message_repository import MessageRepository
-from byoeb.repositories.base_repository import BaseRepository
-from byoeb.chat_app.configuration.config import app_config
+from byoeb.repositories.mongodb_base_repository import MongoBaseRepository
 
-class MongoMessageRepository(MessageRepository, BaseRepository):
+class MongoMessageRepository(MessageRepository, MongoBaseRepository):
     """MongoDB implementation of MessageRepository."""
-
-    def __init__(self, collection_client: AsyncCollection):
-        self._collection = collection_client
-        self._collection_name = app_config["databases"]["mongo_db"]["message_collection"]
-
-    async def find_by_id(self, id: str) -> Optional[Dict[str, Any]]:
-        """Find a single message by its ID."""
-        return await self._collection.find_one({"_id": id})
-
-    async def find_all(self, filter_dict: Optional[Dict[str, Any]] = None, 
-                      projection: Optional[Dict[str, Any]] = None,
-                      sort: Optional[List[Tuple[str, int]]] = None,
-                      limit: Optional[int] = None) -> List[Dict[str, Any]]:
-        """Find multiple messages with optional filtering, projection, sorting, and limiting."""
-        cursor = self._collection.find(filter_dict or {}, projection=projection)
-        if sort:
-            cursor = cursor.sort(sort)
-        if limit is not None:
-            cursor = cursor.limit(limit)
-        return await cursor.to_list(length=None)
-
-    async def count(self, filter_dict: Optional[Dict[str, Any]] = None) -> int:
-        """Count messages matching the filter criteria."""
-        return await self._collection.count_documents(filter_dict or {})
-
-    async def insert_one(self, document: Dict[str, Any]) -> str:
-        """Insert a single message and return its ID."""
-        result = await self._collection.insert_one(document)
-        return str(result.inserted_id)
-
-    async def insert_many(self, documents: List[Dict[str, Any]]) -> List[str]:
-        """Insert multiple messages and return their IDs."""
-        result = await self._collection.insert_many(documents, ordered=False)
-        return [str(doc_id) for doc_id in result.inserted_ids]
-
-    async def update_one(self, filter_dict: Dict[str, Any], 
-                        update_dict: Dict[str, Any]) -> bool:
-        """Update a single message matching the filter criteria."""
-        result = await self._collection.update_one(filter_dict, update_dict)
-        return result.modified_count > 0
-
-    async def update_many(self, filter_dict: Dict[str, Any], 
-                         update_dict: Dict[str, Any]) -> int:
-        """Update multiple messages matching the filter criteria."""
-        result = await self._collection.update_many(filter_dict, update_dict)
-        return result.modified_count
-
-    async def delete_one(self, filter_dict: Dict[str, Any]) -> bool:
-        """Delete a single message matching the filter criteria."""
-        result = await self._collection.delete_one(filter_dict)
-        return result.deleted_count > 0
-
-    async def delete_many(self, filter_dict: Dict[str, Any]) -> int:
-        """Delete multiple messages matching the filter criteria."""
-        result = await self._collection.delete_many(filter_dict)
-        return result.deleted_count
-
-    async def bulk_update(self, bulk_queries: List[Tuple[Dict[str, Any], Dict[str, Any]]]) -> int:
-        """Execute heterogeneous message update queries in bulk."""
-        if not bulk_queries:
-            return 0
-        operations = [UpdateOne(filter=query, update=update) for query, update in bulk_queries]
-        result = await self._collection.bulk_write(operations)
-        return result.modified_count
 
     async def find_messages_by_time_range(self, 
                                         start_timestamp: int,
